pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "videopilot"
version = "0.1.0"
description = "AI-powered video editing & generation platform with 65+ modules"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "VideoPilot" }]

dependencies = [
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "numpy>=1.26",
    "opencv-python-headless>=4.9",
    "Pillow>=10.0",
    "tqdm>=4.66",
    "loguru>=0.7",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
api = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "python-multipart>=0.0.9",
    "aiofiles>=23.0",
]
ml = [
    "torch>=2.2",
    "torchvision>=0.17",
    "transformers>=4.38",
    "diffusers>=0.27",
    "accelerate>=0.27",
    "safetensors>=0.4",
]
audio = [
    "openai-whisper>=20231117",
    "librosa>=0.10",
    "soundfile>=0.12",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "pytest-asyncio>=0.23",
    "ruff>=0.3",
    "mypy>=1.8",
]
all = ["videopilot[api,ml,audio,dev]"]

[project.scripts]
videopilot = "core.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["api", "core", "modules"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.11"
strict = true
